# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AtomQuest Goal Setting & Tracking Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goal-sheets:
    get:
      operationId: listGoalSheets
      tags: [goal-sheets]
      summary: List goal sheets (role-filtered)
      parameters:
        - name: cycleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, approved, returned]
      responses:
        "200":
          description: List of goal sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalSheet"
    post:
      operationId: createGoalSheet
      tags: [goal-sheets]
      summary: Create a goal sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalSheetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
  /goal-sheets/{id}:
    get:
      operationId: getGoalSheet
      tags: [goal-sheets]
      summary: Get goal sheet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal sheet with goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheetDetail"
        "404":
          description: Not found
    patch:
      operationId: updateGoalSheet
      tags: [goal-sheets]
      summary: Update a goal sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalSheetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
  /goal-sheets/{id}/submit:
    post:
      operationId: submitGoalSheet
      tags: [goal-sheets]
      summary: Submit goal sheet for manager approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
  /goal-sheets/{id}/approve:
    post:
      operationId: approveGoalSheet
      tags: [goal-sheets]
      summary: Approve goal sheet (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Approved and locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
  /goal-sheets/{id}/return:
    post:
      operationId: returnGoalSheet
      tags: [goal-sheets]
      summary: Return goal sheet for rework (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnInput"
      responses:
        "200":
          description: Returned for rework
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"
  /goal-sheets/{id}/unlock:
    post:
      operationId: unlockGoalSheet
      tags: [goal-sheets]
      summary: Unlock goal sheet (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockInput"
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSheet"

  # Goals (within sheets)
  /goal-sheets/{sheetId}/goals:
    post:
      operationId: createGoal
      tags: [goals]
      summary: Add a goal to a sheet
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Shared Goals
...
components:
  schemas:
...
    GoalSheet:
      type: object
      required: [id, employeeId, employeeName, cycleId, cycleName, status, totalWeightage, goalsCount]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeDepartment:
          type: string
        cycleId:
          type: integer
        cycleName:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved, returned]
        totalWeightage:
          type: number
        goalsCount:
          type: integer
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        managerComment:
          type: ["string", "null"]
        isLocked:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    GoalSheetDetail:
      type: object
      required: [id, employeeId, employeeName, cycleId, cycleName, status, totalWeightage, goals]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeDepartment:
          type: string
        cycleId:
          type: integer
        cycleName:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved, returned]
        totalWeightage:
          type: number
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        managerComment:
          type: ["string", "null"]
        isLocked:
          type: boolean
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
        createdAt:
          type: string
        updatedAt:
          type: string
    GoalSheetInput:
      type: object
      required: [cycleId]
      properties:
        cycleId:
          type: integer
    GoalSheetUpdate:
      type: object
      properties:
        managerComment:
          type: string
    ApprovalInput:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
    ReturnInput:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
    UnlockInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # Goal
    Goal:
      type: object
      required: [id, goalSheetId, thrustAreaId, thrustAreaName, title, description, uomType, target, weightage, status]
      properties:
        id:
          type: integer
        goalSheetId:
          type: integer
        thrustAreaId:
          type: integer
        thrustAreaName:
          type: string
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: ["string", "null"]
        target:
          type: string
        weightage:
          type: number
        status:
          type: string
          enum: [not_started, on_track, completed]
        isShared:
          type: boolean
        sharedGoalId:
          type: ["integer", "null"]
        isReadOnly:
          type: boolean
        progressScore:
          type: ["number", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [thrustAreaId, title, description, uomType, target, weightage]
      properties:
        thrustAreaId:
          type: integer
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: string
        target:
          type: string
        weightage:
          type: number
    GoalUpdate:
      type: object
      properties:
        thrustAreaId:
          type: integer
        title:
          type: string
        description:
          type: string
        uomType:
          type: string
          enum: [numeric_min, numeric_max, timeline, zero]
        uomUnit:
          type: string
        target:
          type: string
        weightage:
          type: number
        status:
          type: string
          enum: [not_started, on_track, completed]

    # Shared Goals
...